% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export2vcf}
\alias{export2vcf}
\title{Export contacts dataframe to VCF vcard file}
\usage{
export2vcf(contacts_df, filepath_vcf)
}
\arguments{
\item{contacts_df}{cantacts dataframe. Should have FN and at last one of N contact fields}

\item{filepath_vcf}{path to write VCF file}
}
\value{
NA
}
\description{
Export contacts dataframe to VCF vcard file
}
