% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_telegramJSON}
\alias{import_telegramJSON}
\title{Import JSON (from Telegram phone data export)}
\usage{
import_telegramJSON(tlgJSON)
}
\arguments{
\item{tlgJSON}{path to telegram JSON file}
}
\value{
dataframe with Telegram address book
}
\description{
Import JSON (from Telegram phone data export)
}
\examples{
df <- import_telegramJSON("telegram_export.json")
}
