% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_vcf}
\alias{import_vcf}
\title{Import data fram vcf file}
\usage{
import_vcf(filepath)
}
\arguments{
\item{filepath}{path to VCF file}
}
\value{
dataframe
}
\description{
Import data fram vcf file
}
\examples{
\dontrun{
contacts_df <- import_vcf("dir/mycontacts.vcf")
}
}
