% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{prettify_phone_no}
\alias{prettify_phone_no}
\title{Removes leading zeros, spaces in phone number}
\usage{
prettify_phone_no(phone_str)
}
\arguments{
\item{phone_str}{phone number string}
}
\value{
prettified string
}
\description{
Removes leading zeros, spaces in phone number
}
