% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{unify_most_char}
\alias{unify_most_char}
\title{Collapsing vector in one element, which has max str length}
\usage{
unify_most_char(vec)
}
\arguments{
\item{vec}{vector of string parameters or empty string or NA}
}
\value{
str
}
\description{
Collapsing vector in one element, which has max str length
}
