% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2vcard.R
\name{vcard}
\alias{vcard}
\title{Generate one VCARD}
\usage{
vcard(
  name,
  cell = NULL,
  org = NULL,
  email = NULL,
  birthday = NULL,
  address = NULL
)
}
\arguments{
\item{name}{name}

\item{cell}{cell phone number}

\item{org}{company or organization}

\item{email}{email}

\item{birthday}{date of birth (YYYY-MM-DD)}

\item{address}{address}
}
\value{
a VCARD
}
\description{
Generate one VCARD
}
\examples{
data = data.frame(
  name = c("张三","李四"),
  cell = c("13033339999","13800000000"),
  org = c("菜鸟","腾讯")
)
cards = lapply(1:nrow(data), function(i){
  vcard(name = data$name[[i]],
        cell = data$cell[[i]],
        birthday = data$birthday[[i]],
        address = data$address[[i]])
})

con = file("contact.vcf", "a")
lapply(cards, xfun::write_utf8, con = con)
close(con)
}
