% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{vcard_item2df}
\alias{vcard_item2df}
\title{Parses lines of VCF for one contact and makes dataframe for it}
\usage{
vcard_item2df(v_item)
}
\arguments{
\item{v_item}{per contact group of VCF lines encapsulated in dataframe, each line of VCF in row}
}
\value{
dataframe
}
\description{
Parses lines of VCF for one contact and makes dataframe for it
}
