% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_contact}
\alias{write_contact}
\title{Produces VCF contact record for ONE person dataframe rows}
\usage{
write_contact(df)
}
\arguments{
\item{df}{dataframe with ONE person contact fields. From this dataframe will be produced one contact record}
}
\value{
str VCF record for one contact
}
\description{
Produces VCF contact record for ONE person dataframe rows
}
